Correct sign extension for i24 samples
We can store them in 24 bits, but when we later convert to float, we do
need to sign-extend to i32, otherwise it sounds very very bad.

+++ src/playback.rs
@@ -218,7 +218,7 @@ fn write_samples(
                     match block.slice() {
                         SampleDataSlice::I16(src) => {
                             for (i, s) in src.iter().take(n).enumerate() {
-                                let (l, r) = filters.tick_i16(s.0 as i32, s.1 as i32);
+                                let (l, r) = filters.tick_i16(s.0, s.1);
                                 put_i(i, l, r);
                             }
                         }
